Se dividio el main y los usuarios, se arreglo el problema el juego completo.
