+++ ProyectoJuegoJava/ProjectoFinal/Final/src/main/java/Trabajo/Inventario/Pocion.java
@@ -7,12 +7,15 @@ public class Pocion implements Objeto{
     private int peso;
     private boolean esTemporal;
     private TiposDeObjetos tipo;
+    private boolean usado;
+
 
     public Pocion() {
         this.nombre = TiposDeObjetos.POCION.getNombre();
         this.peso = TiposDeObjetos.POCION.getPeso();
         this.esTemporal = TiposDeObjetos.POCION.esTemporal();
         this.tipo = TiposDeObjetos.POCION;
+        this.usado = false;
     }
 
     @Override
@@ -34,6 +37,12 @@ public class Pocion implements Objeto{
     public TiposDeObjetos tipo() {
         return this.tipo;
     }
+
+    @Override
+    public boolean estaUsado() {return usado;}
+
+    @Override
+    public void marcarComoUsado() {this.usado = true;}
 
     @Override
     public void usar(Personaje personaje) {
